+++ src/editor/interface/Background.ts
@@ -1,0 +1,8 @@
+import { BackgroundRepeat, BackgroundSize } from '../dataset/enum/Background'
+
+export interface IBackgroundOption {
+  color?: string
+  image?: string
+  size?: BackgroundSize
+  repeat?: BackgroundRepeat
+}