+++ src/editor/interface/Editor.ts
@@ -5,6 +5,7 @@ import {
   PaperDirection,
   WordBreak
 } from '../dataset/enum/Editor'
+import { IBackgroundOption } from './Background'
 import { ICheckboxOption } from './Checkbox'
 import { IPadding } from './Common'
 import { IControlOption } from './Control'
@@ -29,6 +30,7 @@ export interface IEditorData {
 export interface IEditorOption {
   mode?: EditorMode
   defaultType?: string
+  defaultColor?: string
   defaultFont?: string
   defaultSize?: number
   minSize?: number
@@ -40,7 +42,6 @@ export interface IEditorOption {
   height?: number
   scale?: number
   pageGap?: number
-  backgroundColor?: string
   underlineColor?: string
   strikeoutColor?: string
   rangeColor?: string
@@ -81,6 +82,7 @@ export interface IEditorOption {
   group?: IGroup
   pageBreak?: IPageBreak
   zone?: IZoneOption
+  background?: IBackgroundOption
 }
 
 export interface IEditorResult {